+++ packages/terra-search-field/tests/nightwatch/SearchFieldTests.jsx
@@ -6,11 +6,11 @@ import { Link } from 'react-router';
 const SearchFieldTests = () => (
   <div>
     <ul>
-      <li><Link to="/tests/search-field-tests/default">Default SearchField</Link></li>
-      <li><Link to="/tests/search-field-tests/placeholder">Placeholder SearchField</Link></li>
-      <li><Link to="/tests/search-field-tests/callback">Callback SearchField</Link></li>
-      <li><Link to="/tests/search-field-tests/delayed">1000ms Delayed SearchField</Link></li>
-      <li><Link to="/tests/search-field-tests/minimum-length">5 Character Minimum Length SearchField</Link></li>
+      <li><Link to="/tests/search-field-tests/default">SearchField - Default</Link></li>
+      <li><Link to="/tests/search-field-tests/placeholder">SearchField - Placeholder</Link></li>
+      <li><Link to="/tests/search-field-tests/callback">SearchField - Callback for Valid and Invalid Entry</Link></li>
+      <li><Link to="/tests/search-field-tests/delayed">SearchField - 1000ms Delayed </Link></li>
+      <li><Link to="/tests/search-field-tests/minimum-length">SearchField - 5 Character Minimum Length </Link></li>
     </ul>
   </div>
 );